requirements.txt
reporoulette
huggingface_hub
pandas
python-dotenv
selenium 
beautifulsoup4